Use ref and out keywords to pass structs like FVector in the C# wrapper
This isn't actually limited to just FVector, any of the basic struct
proxies like FQuat, FTransform etc. and primitive types will work.

For example, if a UFunction is declared with an "FVector&" parameter the
corresponding method in the C# wrapper class will have an "out FVector"
parameter. "ref FVector" may seem like a better choice for "FVector&"
but the generator lets UHT determine what is an out and what is a ref
parameter, and the code generated thus far appears to make sense as is.

Fixes #2