Adjust build settings for Klawr.ClrHost.Native lib
The build settings now more closely resemble those used for Box2D,
under the assumption that the Box2D build settings closely resemble
(or are at least compatible with) whatever concoction of build settings
UBT uses to build the engine and game projects.

Will need to update the Win32 platform settings to match the Win64
platform settings in the future.