Workaround for missing GeneratedScriptLibraries.inl in ScriptPlugin
I've disabled the ScriptGeneratorPlugin in my engine source since
it just wastes time, but this broke the ScriptPlugin build because the
file with the generated wrapper code wasn't being generated anymore. For
now I've just commented out the include that references the missing
generated file, but obviously that's not a real solution, will come up
with something better later unless one of the other people working on
the various script integrations beat me to it.