Add a custom C# wrapper class for UKlawrScriptComponent
The C# UKlawrScriptComponent class is currently an abstract class, the
idea is that users will subclass it to create script components
that can be accessed in Blueprints via the C++ UKlawrScriptComponent
class.

Currently the native UKlawrScriptComponent is hard-coded to create
a managed TestComponent instance for testing purposes, but this will
obviously be configurable in the future.