Reload the game scripts assembly after the user creates a new C# script
When a new Klawr Blueprint is created at the user's behest a new C#
source file is also added to the game scripts assembly project. The
assembly is then rebuilt, and if the build succeeds the assembly is
reloaded. However, assemblies can't actually be unloaded without
unloading the app domain they were loaded into initially, so what really
end happens is the app domain is reloaded, and then finally the new
Blueprint is actually created.