Auto-rebuild and auto-reload the game scripts assembly
The editor plugin now monitors the script source directories for
changes, if the user adds or modifies a script file the game scripts
assembly project will be automatically updated (if the user added
files), rebuilt, and finally the assembly will be reloaded if the build
succeeds.

The editor plugin also monitors the game scripts assembly itself, and
reloads it if the user rebuilds the assembly outside UnrealEd, or
replaces it in any other way.