Rename the generated game scripts project
The generated C# project for the game scripts assembly is now named
<ProjectName>Scripts.csproj and is located in
Intermediate/ProjectFiles/<ProjectName>Scripts/. This change was made to
avoid name confusion between the C++ project and the C# project files
which up to this point only differed by their extension.

The name of the game scripts assembly itself remains unchanged.