Change the editor plugin type from Editor to Developer
To quote Epic's Michael Noland:
"Developer means it loads in WITH_EDITOR builds (e.g., -game builds),
not just when actually running the editor, and the loading phase
controls when it gets loaded. For anything related to Blueprint or
Anim BP nodes, you'll want Developer as opposed to Editor, since we
still compile-on-run in -game.

The Editor type is only for actual interactive GUI stuff that is only
shown when running the editor (which could also live in a Developer
module)."