Marshal `FName` arguments as `FScriptName` instead of `const TCHAR*`
This fixes wrapper generation for UPrimitiveComponent::K2_LineTraceComponent()
which takes an FName argument by reference. The downside of this change
is that currently there's no way to convert an FScriptName to a managed
string or vice versa, but this should be relatively straightforward to
implement... can be as simple as a couple of utility functions to do the
conversions.