Stop the Klawr code generator from processing the Klawr runtime plugin
The UKlawrScriptComponent class in the runtime plugin already has
a managed counterpart defined in the template project used to build the
managed wrappers, so the code generator shouldn't auto-generate another
wrapper for it. For now it's sufficient to disable wrapper generation
for the entire runtime plugin to work around the problem.