Fix Klawr Build.cs files for UE 4.8
UnrealBuildTool.BuildingRocket() was replaced by IsEngineInstalled().
The default script plugins no longer care whether they're built with the
Launcher/Rocket or GitHub version of the engine (as of four months ago),
so I guess the Klawr plugins no longer need to ensure they're being
built with the GitHub version of the engine (though I haven't yet tried
building them with the Launcher version of the engine).